refactor(server): Extract `runScraper` util
